requirements.txt
feedparser>=6.0.0
deep-translator>=1.11.0
python-dotenv>=1.0.0
pytz
